# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Court Session Management AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthUser {
  username: string;
}

export interface DashboardStats {
  totalCases: number;
  todayHearings: number;
  upcomingHearings: number;
  finishedHearings: number;
}

export interface MessageInput {
  /**
     * Raw court SMS text pasted by the secretary
     * @minLength 1
     */
  message: string;
}

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  Upcoming: 'Upcoming',
  Today: 'Today',
  Finished: 'Finished',
  Cancelled: 'Cancelled',
} as const;

/**
 * Fields extracted by AI from the SMS message. Any missing field is null.
 */
export interface SessionExtraction {
  /** @nullable */
  case_number: string | null;
  /** @nullable */
  plaintiff: string | null;
  /** @nullable */
  defendant: string | null;
  /** @nullable */
  court: string | null;
  /** @nullable */
  court_circuit: string | null;
  /** @nullable */
  case_subject: string | null;
  /** @nullable */
  session_type: string | null;
  /** @nullable */
  session_date_hijri: string | null;
  /** @nullable */
  session_time: string | null;
  /** @nullable */
  notes: string | null;
}

/**
 * Reviewed/edited fields submitted to create a new session row
 */
export interface SessionInput {
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  plaintiff?: string | null;
  /** @nullable */
  defendant?: string | null;
  /** @nullable */
  court?: string | null;
  /** @nullable */
  courtCircuit?: string | null;
  /** @nullable */
  caseSubject?: string | null;
  /** @nullable */
  sessionType?: string | null;
  /** @nullable */
  sessionDateHijri?: string | null;
  /** @nullable */
  sessionTime?: string | null;
  /** @nullable */
  notes?: string | null;
}

/**
 * Partial update to an existing session row
 */
export interface SessionUpdate {
  /** @nullable */
  caseNumber?: string | null;
  /** @nullable */
  plaintiff?: string | null;
  /** @nullable */
  defendant?: string | null;
  /** @nullable */
  court?: string | null;
  /** @nullable */
  courtCircuit?: string | null;
  /** @nullable */
  caseSubject?: string | null;
  /** @nullable */
  sessionType?: string | null;
  /** @nullable */
  sessionDateHijri?: string | null;
  /** @nullable */
  sessionTime?: string | null;
  /** @nullable */
  notes?: string | null;
  status?: SessionStatus;
}

export interface Session {
  /** 1-based row index within the Google Sheet */
  id: number;
  /** @nullable */
  caseNumber: string | null;
  /** @nullable */
  plaintiff: string | null;
  /** @nullable */
  defendant: string | null;
  /** @nullable */
  court: string | null;
  /** @nullable */
  courtCircuit: string | null;
  /** @nullable */
  caseSubject: string | null;
  /** @nullable */
  sessionType: string | null;
  /** @nullable */
  sessionDateHijri: string | null;
  /** @nullable */
  sessionTime: string | null;
  /** @nullable */
  notes: string | null;
  status: SessionStatus;
  reminder24: boolean;
  reminder6: boolean;
  createdAt: string;
}

export type ListSessionsParams = {
status?: SessionStatus;
};