# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Court Session Management AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /ai/analyze:
    post:
      operationId: analyzeMessage
      tags: [ai]
      summary: Analyze a pasted court SMS message with AI and extract structured fields
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: Extracted session fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionExtraction"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: AI extraction failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
    AuthUser:
      type: object
      required: [username]
      properties:
        username:
          type: string
...
    MessageInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          description: Raw court SMS text pasted by the secretary
...
    SessionExtraction:
      type: object
      description: Fields extracted by AI from the SMS message. Any missing field is null.
      required:
        [
          case_number,
          plaintiff,
          defendant,
          court,
          court_circuit,
          case_subject,
          session_type,
          session_date_hijri,
          session_time,
          notes,
        ]
      properties:
        case_number:
          type: ["string", "null"]
        plaintiff:
          type: ["string", "null"]
        defendant:
          type: ["string", "null"]
        court:
          type: ["string", "null"]
        court_circuit:
          type: ["string", "null"]
        case_subject:
          type: ["string", "null"]
        session_type:
          type: ["string", "null"]
        session_date_hijri:
          type: ["string", "null"]
        session_time:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...