# lib/api-zod/src/generated/types/sessionExtraction.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Court Session Management AI API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Fields extracted by AI from the SMS message. Any missing field is null.
 */
export interface SessionExtraction {
  /** @nullable */
  case_number: string | null;
  /** @nullable */
  plaintiff: string | null;
  /** @nullable */
  defendant: string | null;
  /** @nullable */
  court: string | null;
  /** @nullable */
  court_circuit: string | null;
  /** @nullable */
  case_subject: string | null;
  /** @nullable */
  session_type: string | null;
  /** @nullable */
  session_date_hijri: string | null;
  /** @nullable */
  session_time: string | null;
  /** @nullable */
  notes: string | null;
}